refactor(semantic-search): extract production wiring out of main.ts
The ~430-line semantic-search wiring block lived inside onload() with no
test coverage, including the silent stale-store wipe (the riskiest
startup path). Move it to services/productionWiring.ts as
wireSemanticSearch(plugin), which onload now awaits and assigns. The
probe plus silent-wipe logic is split into an exported
probeAndWipeStaleStores helper so it is unit-testable with an in-memory
adapter. The construction-time semanticSearch read drops its
globalSettingsMutex wrapper for SettingsStore.readSlice. main.ts falls
from 604 to 141 lines and loses ten deep-imports of semantic-search
internals. Adds 4 wipe-path tests; the compiled bundle is byte-identical
to before, confirming behavior preservation.
